Move secrets to env vars
