+++ .gitignore
@@ -1,6 +1,3 @@
 node_modules
-config/secrets.json
+config/.env
 .idea
-client-secret.json
-api_key
-credentials.tar.gz
