+++ config/.env.example
@@ -1,0 +1,4 @@
+DISCORD_TOKEN=roughly 60 character token
+DISCORD_CLIENTID=123456789012345678
+WEBHOOK_ID=
+WEBHOOK_TOKEN=
