+++ config/README.md
@@ -1,0 +1,16 @@
+# Setting up secrets
+
+Create a `.env` file in this directory with the following structure:
+```
+DISCORD_TOKEN=
+DISCORD_CLIENTID=
+WEBHOOK_ID=
+WEBHOOK_TOKEN=
+```
+
+Webhook example:
+```
+https://discord.com/api/webhooks/<WEBHOOK_ID>/<WEBHOOK_TOKEN>
+```
+The discord credentials are required.
+Webhook is for debugging to a discord channel.
