+++ package.json
@@ -8,6 +8,7 @@
     "bunyan": "^1.8.12",
     "cors": "^2.8.4",
     "discord.js": "^11.3.2",
+    "dotenv": "^8.2.0",
     "express": "^4.16.2",
     "mocha": "^5.2.0",
     "morgan": "^1.9.1"
